Refactor Indexing data.
